requirements.txt
pygame==2.5.2
pyinstaller==6.3.0
pillow==10.2.0 